+++ Doremi/Core/Include/EntityComponent/Components/AudioComponent.hpp
@@ -8,8 +8,8 @@ namespace Doremi
         */
         struct AudioComponent
         {
-            size_t soundID;
-            AudioComponent(size_t p_soundID) { soundID = p_soundID; }
+            int soundID;
+            AudioComponent(int p_soundID) { soundID = p_soundID; }
             AudioComponent() : soundID(-1) {}
         };
     }
